+++ packages/compass-aggregations/src/utils/configure-store.js
