+++ packages/compass-aggregations/src/utils/get-directory.js
